+++ TidalPy/Tides_x/potential/global_.hpp
@@ -1,0 +1,19 @@
+#pragma once
+
+#include "obliquity_driver_.hpp"
+#include "eccentricity_driver_.hpp"
+
+
+struct GlobalPotentialResult
+{
+    c_IntMap<c_Key4, double> modes;
+    c_IntMap<c_IntMap<
+    int error_code = 0;
+};
+
+
+GlobalPotentialResult global_potential()
+{
+
+
+}