+++ TidalPy/rheology/models.pyx
@@ -4,7 +4,11 @@
 from libc.math cimport fabs, cos, sin, tgamma, isinf
 
 from TidalPy.exceptions import UnknownModelError
-from TidalPy.constants cimport d_MIN_FREQUENCY, d_MAX_FREQUENCY, d_MIN_MODULUS, d_PI_DBL, d_INF_DBL
+from TidalPy.constants cimport TidalPyConfig, d_PI, d_INF, get_shared_config_address, tidalpy_config_ptr
+
+# Wire up the pointer at import time
+if tidalpy_config_ptr == NULL:
+    tidalpy_config_ptr = get_shared_config_address()
 from TidalPy.utilities.math.complex cimport cf_build_dblcmplx
 
 
@@ -72,11 +76,11 @@ cdef class Newton(RheologyModelBase):
         frequency_abs = fabs(frequency)
 
         # Check for extreme values. If found: use pre-calculated limits.
-        if frequency_abs < d_MIN_FREQUENCY:
+        if frequency_abs < tidalpy_config_ptr.d_MIN_FREQUENCY:
             return cf_build_dblcmplx(0.0, 0.0)
-        elif frequency_abs > d_MAX_FREQUENCY or isinf(frequency_abs):
-            return cf_build_dblcmplx(0.0, d_INF_DBL)
-        if modulus < d_MIN_MODULUS:
+        elif frequency_abs > tidalpy_config_ptr.d_MAX_FREQUENCY or isinf(frequency_abs):
+            return cf_build_dblcmplx(0.0, d_INF)
+        if modulus < tidalpy_config_ptr.d_MIN_MODULUS:
             return cf_build_dblcmplx(0.0, frequency_abs * viscosity)
 
         return cf_build_dblcmplx(0.0, frequency_abs * viscosity)
@@ -97,11 +101,11 @@ cdef class Maxwell(RheologyModelBase):
         cdef double frequency_abs = fabs(frequency)
 
         # Check for extreme values. If found: use pre-calculated limits.
-        if frequency_abs < d_MIN_FREQUENCY:
+        if frequency_abs < tidalpy_config_ptr.d_MIN_FREQUENCY:
             return cf_build_dblcmplx(0.0, 0.0)
-        elif frequency_abs > d_MAX_FREQUENCY or isinf(frequency_abs):
+        elif frequency_abs > tidalpy_config_ptr.d_MAX_FREQUENCY or isinf(frequency_abs):
             return cf_build_dblcmplx(modulus, 0.0)
-        if modulus < d_MIN_MODULUS:
+        if modulus < tidalpy_config_ptr.d_MIN_MODULUS:
             return cf_build_dblcmplx(0.0, 0.0)
 
         cdef double maxwell_time  = viscosity / modulus
@@ -134,11 +138,11 @@ cdef class Voigt(RheologyModelBase):
         cdef double voigt_visosity = self.voigt_viscosity_scale * viscosity
 
         # Check for extreme values. If found: use pre-calculated limits.
-        if frequency_abs < d_MIN_FREQUENCY:
+        if frequency_abs < tidalpy_config_ptr.d_MIN_FREQUENCY:
             return cf_build_dblcmplx(voigt_modulus, 0.0)
-        elif frequency_abs > d_MAX_FREQUENCY or isinf(frequency_abs):
-            return cf_build_dblcmplx(0.0, d_INF_DBL)
-        if modulus < d_MIN_MODULUS:
+        elif frequency_abs > tidalpy_config_ptr.d_MAX_FREQUENCY or isinf(frequency_abs):
+            return cf_build_dblcmplx(0.0, d_INF)
+        if modulus < tidalpy_config_ptr.d_MIN_MODULUS:
             return cf_build_dblcmplx(0.0, voigt_visosity * frequency_abs)
 
         return cf_build_dblcmplx(voigt_modulus, frequency_abs * voigt_visosity)
@@ -166,11 +170,11 @@ cdef class Burgers(RheologyModelBase):
         cdef double frequency_abs = fabs(frequency)
 
         # Check for extreme values. If found: use pre-calculated limits.
-        if frequency_abs < d_MIN_FREQUENCY:
+        if frequency_abs < tidalpy_config_ptr.d_MIN_FREQUENCY:
             return cf_build_dblcmplx(0.0, 0.0)
-        elif frequency_abs > d_MAX_FREQUENCY or isinf(frequency_abs):
+        elif frequency_abs > tidalpy_config_ptr.d_MAX_FREQUENCY or isinf(frequency_abs):
             return cf_build_dblcmplx(modulus, 0.0)
-        if modulus < d_MIN_MODULUS:
+        if modulus < tidalpy_config_ptr.d_MIN_MODULUS:
             return cf_build_dblcmplx(0.0, 0.0)
 
         cdef double voigt_modulus       = self.voigt_modulus_scale * modulus
@@ -202,7 +206,7 @@ cdef class Andrade(RheologyModelBase):
         self.alpha           = new_args[0]
         self.zeta            = new_args[1]
         self.alpha_factorial = tgamma(self.alpha + 1.)
-        self.sine_term       = cos(d_PI_DBL * self.alpha / 2.) - 1.0j * sin(d_PI_DBL * self.alpha / 2.)
+        self.sine_term       = cos(d_PI * self.alpha / 2.) - 1.0j * sin(d_PI * self.alpha / 2.)
 
     cdef double complex _implementation(
             self,
@@ -214,11 +218,11 @@ cdef class Andrade(RheologyModelBase):
         cdef double frequency_abs = fabs(frequency)
 
         # Check for extreme values. If found: use pre-calculated limits.
-        if frequency_abs < d_MIN_FREQUENCY:
+        if frequency_abs < tidalpy_config_ptr.d_MIN_FREQUENCY:
             return cf_build_dblcmplx(0.0, 0.0)
-        elif frequency_abs > d_MAX_FREQUENCY or isinf(frequency_abs):
+        elif frequency_abs > tidalpy_config_ptr.d_MAX_FREQUENCY or isinf(frequency_abs):
             return cf_build_dblcmplx(modulus, 0.0)
-        if modulus < d_MIN_MODULUS:
+        if modulus < tidalpy_config_ptr.d_MIN_MODULUS:
             return cf_build_dblcmplx(0.0, 0.0)
 
         cdef double maxwell_time = viscosity / modulus
@@ -246,7 +250,7 @@ cdef class SundbergCooper(RheologyModelBase):
         self.alpha                 = new_args[2]
         self.zeta                  = new_args[3]
         self.alpha_factorial       = tgamma(self.alpha + 1.)
-        self.sine_term             = cos(d_PI_DBL * self.alpha / 2.) - 1.0j * sin(d_PI_DBL * self.alpha / 2.)
+        self.sine_term             = cos(d_PI * self.alpha / 2.) - 1.0j * sin(d_PI * self.alpha / 2.)
 
     cdef double complex _implementation(
             self,
@@ -258,11 +262,11 @@ cdef class SundbergCooper(RheologyModelBase):
         cdef double frequency_abs = fabs(frequency)
 
         # Check for extreme values. If found: use pre-calculated limits.
-        if frequency_abs < d_MIN_FREQUENCY:
+        if frequency_abs < tidalpy_config_ptr.d_MIN_FREQUENCY:
             return cf_build_dblcmplx(0.0, 0.0)
-        elif frequency_abs > d_MAX_FREQUENCY or isinf(frequency_abs):
+        elif frequency_abs > tidalpy_config_ptr.d_MAX_FREQUENCY or isinf(frequency_abs):
             return cf_build_dblcmplx(modulus, 0.0)
-        if modulus < d_MIN_MODULUS:
+        if modulus < tidalpy_config_ptr.d_MIN_MODULUS:
             return cf_build_dblcmplx(0.0, 0.0)
 
         cdef double voigt_modulus       = self.voigt_modulus_scale * modulus