+++ TidalPy/tides/multilayer/heating.pyx
@@ -19,7 +19,11 @@ cimport numpy as cnp
 cnp.import_array()
 
 from TidalPy.exceptions import ArgumentException
-from TidalPy.constants cimport d_G, d_NAN_DBL
+from TidalPy.constants cimport d_NAN, TidalPyConfig, get_shared_config_address, tidalpy_config_ptr
+
+# Wire up the pointer at import time
+if tidalpy_config_ptr == NULL:
+    tidalpy_config_ptr = get_shared_config_address()
 from TidalPy.RadialSolver cimport RadialSolverSolution
 from TidalPy.Material.eos.eos_solution cimport EOSSolutionCC
 from TidalPy.tides.multilayer.sensitivity cimport cf_calc_sensitivity_to_shear, cf_calc_sensitivity_to_bulk
@@ -104,7 +108,7 @@ def calc_radial_volumetric_tidal_heating(
         double[::1] radial_sensitivity_to_bulk_arr,
         double complex[::1] complex_bulk_modulus_arr,
         int degree_l = 2,
-        double G_to_use = d_G,
+        double G_to_use = tidalpy_config_ptr.d_G,
         cpp_bool perform_checks = True
         ):
 
@@ -231,7 +235,7 @@ def calc_radial_volumetric_tidal_heating_from_rs_solution(
         radial_sensitivity_to_bulk_ptr,
         bulk_modulus_array_ptr,
         degree_l,
-        d_G
+        tidalpy_config_ptr.d_G
         )
     
     return radial_volumetric_tidal_heating_arr