+++ TidalPy/utilities/math/special.pyx
@@ -3,7 +3,8 @@
 
 from libc.math cimport tgamma
 
-from TidalPy.constants cimport d_NAN_DBL
+from TidalPy.constants cimport d_NAN
+
 
 cdef double[51] pre_calculated_doubles
 cdef double* pre_calculated_doubles_ptr = &pre_calculated_doubles[0]
@@ -70,7 +71,7 @@ cdef double cf_double_factorial(int n) noexcept nogil:
         return tgamma(<double>n + 1.) / cf_double_factorial(n - 1)
     else:
         # ValueError('C function `tgamma` experiences overflow for l > 170.')
-        return d_NAN_DBL
+        return d_NAN
 
 
 def double_factorial(int n):